add resume section
This splits out the core of the app into a `pages` section with a
separate layout so that I don't have to turn stuff off in the print
segment.

This is a bit of an early generalization for a single printable page,
but if there is no other way to set the layout of a folder as the root
layout this is likely a wise choice in general. All alternate
substantial layout changes will end up having this problem. This is a
bit of an awkward oversight in Next.js which will likely get tightened
up later on as they add more conveniences to the app router.