fix hydration error in TimeStamp
The github action was running with TZ=UTC (or something equivalent) so I
was seeing a hydration error only in production. I added an explicit
fallback timezone setting for this rather than updating the timezone
things use on the server because I would rather that things like log
timestamps remain unchanged.

This feature is just trying to get a nice compromise in terms of flicker
and correctness for users.