fix source maps for mdx files
This adds in a route to serve source files so that file:// urls don't
need to be relied on. In the process of building this I've found out
that it is somehow just not possible to do something like this and have
it only work in development. It boggles the mind.

There isn't even a way to say that you just don't care about all the
things that can't be generated with `output: export`. Though that aligns
with the fact that linting is built into the compile by default, which
is simply the wrong choice. Anything that absolutely doesn't have to
break the build, is only producing bugs if it chooses to break the build
anyway. A team that is actually going to leverage checks like that
effectively has the skills to set them up in the appropriate locations.