Restructure resume to be parseable
Many tools are not smart enough to handle an unlabelled experience
section correctly, even when they are capable of deducing that the
unlabelled content is obviously job experience because this is a resume.

This results in things like trying to read the first company name as a
section header and leaving everything misaligned by one row.